Added solarized8 color scheme v1.3.0
v1.4.0 caused vimtex imaps to stop working inline, and ultisnip snippets to stop working inline, and matching of the delimiter $ to be incorrect - it always matched with the previous $ even if the matching one was ahead. All of this because of the color scheme! 
I tried uninstalling and reinstalling every plugin, downgrading plugin and neovim versions, and removing one plugin at a time. But finally I tried using a minimal.vim file and added a line at a time when I finally found out it was caused by the colorscheme.